Splits keuzevelden.class.php
