+++ lib/view/formulier/keuzevelden/RequiredColorField.class.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * RequiredColorField.class.php
+ *
+ * @author Jan Pieter Waagmeester <jieter@jpwaag.com>
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see ColorField
+ */
+class RequiredColorField extends ColorField {
+
+	public $required = true;
+
+}
