+++ lib/view/formulier/keuzevelden/RequiredGeslachtField.class.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * RequiredGeslachtField.class.php
+ *
+ * @author Jan Pieter Waagmeester <jieter@jpwaag.com>
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see GeslachtField
+ */
+class RequiredGeslachtField extends GeslachtField {
+
+	public $required = true;
+
+}
