+++ lib/view/formulier/keuzevelden/RequiredKerkField.class.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * RequiredKerkField.class.php
+ *
+ * @author Jan Pieter Waagmeester <jieter@jpwaag.com>
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see KerkField
+ */
+class RequiredKerkField extends KerkField {
+
+	public $required = true;
+
+}
