+++ lib/view/formulier/keuzevelden/RequiredTimeField.class.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * RequiredTimeField.class.php
+ *
+ * @author Jan Pieter Waagmeester <jieter@jpwaag.com>
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see TimeField
+ */
+class RequiredTimeField extends TimeField {
+
+	public $required = true;
+
+}
