+++ lib/view/formulier/keuzevelden/RequiredWeekdagField.class.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * RequiredWeekdagField.class.php
+ *
+ * @author Jan Pieter Waagmeester <jieter@jpwaag.com>
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see WeekdagField
+ */
+class RequiredWeekdagField extends WeekdagField {
+
+	public $required = true;
+
+}
